feature: make nmeasParser ts compatible
