--- serial-input/src/components/nmeaParser/index.js
@@ -1,1 +1,0 @@
-export { default } from './NMEAParser.js';
