+++ serial-input/src/components/nmeaParser/index.ts
@@ -1,0 +1,1 @@
+export { default } from './NMEAParser';
